% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_building.R
\name{actions_jump_always}
\alias{actions_jump_always}
\title{Create a "jump always" action for a logic field}
\usage{
actions_jump_always(to)
}
\arguments{
\item{to}{the ref to jump to}
}
\description{
Create a "jump always" action for a logic field
}
