% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_building.R
\name{create_form}
\alias{create_form}
\title{Create a Typeform}
\usage{
create_form(form_object, auth = config::get()$typeform)
}
\arguments{
\item{form_object}{the form object to provide the forms API endpoint}

\item{auth}{personal access token}
}
\value{
if error, response. If success, the link to the live form
}
\description{
Create a Typeform
}
\examples{
new_form <- form_object(title = "New Form") |>
  form_field(
    title = "Select a Color",
    type = "multiple_choice",
    ref = "color_select",
    properties = field_prop_multi(
      description = "Select a color form the given choices",
      labels = c("Red", "Blue", "Green")
    )
  )

link <- crate_form(new_form)
browseURL(link)
}
