% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhooks.R
\name{create_webhook}
\alias{create_webhook}
\title{Create or Update a Webhook}
\usage{
create_webhook(form_id, tag, url, ..., auth = config::get()$typeform)
}
\arguments{
\item{form_id}{the form id, can be found in url}

\item{tag}{a tag for the webhook. Required.}

\item{url}{url to point the webhook to.}

\item{...}{see details}

\item{auth}{auth}
}
\description{
Create or Update a Webhook
}
\details{
enabled: boolean
True if you want to send responses to the webhook immediately. Otherwise, false.
secret: string
If specified, will be used to sign the webhook payload with HMAC SHA256, so that you can verify that it came from Typeform.
url: string
Webhook URL.
verify_ssl: boolean
True if you want Typeform to verify SSL certificates when delivering payloads.
}
