% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_mgmt.R
\name{delete_forms}
\alias{delete_forms}
\title{Delete a Form}
\usage{
delete_forms(form_vector, auth = config::get()$typeform)
}
\arguments{
\item{form_vector}{a character vector of form id's. Canbe single}

\item{auth}{personal access token}
}
\description{
Delete a Form
}
