% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_building.R
\name{field_multiple_choice_properties}
\alias{field_multiple_choice_properties}
\title{Create properties for a multiple choice field.}
\usage{
field_multiple_choice_properties(
  description,
  labels,
  refs = snakecase::to_snake_case(labels),
  allow_multiple_selection = FALSE,
  randomize = FALSE,
  allow_other_choice = FALSE,
  vertical_alignment = TRUE
)
}
\arguments{
\item{description}{The instructions for the question}

\item{labels}{human readable option labels}

\item{refs}{machine readable option labels}

\item{allow_multiple_selection}{boolean}

\item{randomize}{boolean}

\item{allow_other_choice}{boolean}

\item{vertical_alignment}{true to list answer choices vertically. false to list answer choices horizontally.}
}
\value{
returns a properties object to be used in creating a field
}
\description{
Supplied to the \code{properties} argument in \code{type_create_form_field}.
}
