% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_building_fields.R
\name{form_field_hidden}
\alias{form_field_hidden}
\title{Create hidden fields}
\usage{
form_field_hidden(form_object, fields, verbose = TRUE)
}
\arguments{
\item{form_object}{a form created by type_create_form_object.}

\item{fields}{a character vector of field names}

\item{verbose}{should a pretty printed json be shown after each call?}
}
\description{
Create hidden fields
}
