% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_building.R
\name{form_object}
\alias{form_object}
\title{Create a Form Object}
\usage{
form_object(title = paste("New Form", Sys.Date()))
}
\arguments{
\item{title}{Title of the form}
}
\description{
Creates a list that can be appended with form fields, attachments, etc.
}
\details{
reference: https://developer.typeform.com/create/reference/create-form/
}
