% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_building.R
\name{form_settings}
\alias{form_settings}
\title{Create settings for a form}
\usage{
form_settings(form_object, settings = list(hide_navigation = FALSE))
}
\arguments{
\item{form_object}{the form object}

\item{settings}{a named list of settings.}
}
\description{
Create settings for a form
}
