% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_mgmt.R
\name{get_forms}
\alias{get_forms}
\title{Title}
\usage{
get_forms(..., auth = config::get()$typeform)
}
\arguments{
\item{...}{query parameters. defaults to page_size = 200. See details}

\item{auth}{}
}
\description{
Title
}
\details{
query parameters are
search: a string to search for
page
pagesize
workspace_id
sort_by
order_by
}
