% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhooks.R
\name{get_webhooks}
\alias{get_webhooks}
\title{Get a list of webhooks}
\usage{
get_webhooks(form_id, auth = config::get()$typeform)
}
\arguments{
\item{form_id}{form id, can be foundin URL}

\item{auth}{auth}
}
\description{
Get a list of webhooks
}
