% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_building.R
\name{jump_actions}
\alias{jump_actions}
\title{Create vector of actions for a logic jump ref}
\usage{
jump_actions(ref, to, operation, op_value, monitored_type)
}
\arguments{
\item{ref}{the reference being checked in the condition. Often the same ref as the logic field}

\item{to}{the ref to jump to}

\item{operation}{the operation to be applied}

\item{op_value}{the value being tested in the operation}

\item{monitored_type}{the type of field, "hidden" or "field" being checked in the operation}
}
\value{
returns a field that can be nested in the actions array of a logic jump.
}
\description{
Create vector of actions for a logic jump ref
}
