% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_building.R
\name{logic_field}
\alias{logic_field}
\title{Create a logic field}
\usage{
logic_field(form_object, ref, ref_type, actions)
}
\arguments{
\item{form_object}{a form object}

\item{ref}{the field for which a logical condition is checked}

\item{ref_type}{either "field" or "hidden"}

\item{actions}{an array of actions and conditions. Caan be created with functions like \code{jump_actions()}.}
}
\description{
Create a logic field
}
