% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_building_fields.R
\name{thank_you_screens}
\alias{thank_you_screens}
\title{Thank You Screens for Ending}
\usage{
thank_you_screens(
  form_object,
  title = "Response Recorded",
  ref = snakecase::to_snake_case(title),
  type = "thankyou_screen",
  show_button = TRUE,
  button_text = "Record Again",
  button_mode = "reload",
  share_icons = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{form_object}{a form created by type_create_form_object.}

\item{title}{the field title.}

\item{ref}{the machine readable name for the referenced field.}

\item{type}{Valid values: "thankyou_screen", "url_redirect". The type of thank you screen.}

\item{show_button}{T/F display button}

\item{button_text}{button text}

\item{button_mode}{Valid values: "reload", "default_redirect", "redirect"}

\item{verbose}{should a pretty printed json be shown after each call?}
}
\description{
Thank You Screens for Ending
}
