% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_building.R
\name{type_create_form_object}
\alias{type_create_form_object}
\title{Create a Form Object}
\usage{
type_create_form_object(title = paste("New Form", Sys.Date()))
}
\arguments{
\item{title}{Title of the form}
}
\description{
Creates a list that can be appended with form fields, attachments, etc.
}
