% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{type_request}
\alias{type_request}
\title{Send a request to the typeform API}
\usage{
type_request(endpoint, auth = config::get()$typeform)
}
\arguments{
\item{endpoint}{API endpoint}

\item{auth}{auth token. Defaults to a config.yml with a default => typeform entry}
}
\description{
Send a request to the typeform API
}
