% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_responses.R
\name{type_responses}
\alias{type_responses}
\title{Retrieve Responses from a Typeform}
\usage{
type_responses(
  form_id,
  ...,
  attempt_table = TRUE,
  wide_table = TRUE,
  auth = config::get()$typeform
)
}
\arguments{
\item{form_id}{the form id, can be found in the URL of a form}

\item{...}{query paramters. See details}

\item{attempt_table}{coerce the results to a table?}

\item{wide_table}{only if attempt_table = TRUE, simplify table results to a wide format}

\item{auth}{typeform API token}
}
\description{
Retrieve Responses from a Typeform
}
