% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhooks.R
\name{update_webhook}
\alias{update_webhook}
\title{Create or Update a Webhook}
\usage{
update_webhook(form_id, tag, url, ..., auth = config::get()$typeform)
}
\arguments{
\item{form_id}{the form id, can be found in url}

\item{tag}{a tag for the webhook. Required.}

\item{url}{url to point the webhook to.}

\item{...}{see details}

\item{auth}{auth}
}
\description{
Create or Update a Webhook
}
