% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_responses.R
\name{val_extractor}
\alias{val_extractor}
\title{function factory that generates an extractor function for a given type}
\usage{
val_extractor(type)
}
\arguments{
\item{type}{the type, given in the answers object.}
}
\description{
function factory that generates an extractor function for a given type
}
