% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_building.R
\name{form_field}
\alias{form_field}
\title{Create a form field, often a question in the form.}
\usage{
form_field(
  form_object,
  title,
  type,
  ref = snakecase::to_snake_case(title),
  properties = NULL,
  validations = NULL,
  attachment = NULL,
  layout = NULL,
  verbose = FALSE
)
}
\arguments{
\item{form_object}{a form created by type_create_form_object.}

\item{title}{the field title.}

\item{type}{type of field, See details for more.}

\item{ref}{the machine readable name for the referenced field.}

\item{properties}{field properties. Can be created with field_prop_multi or other appropraite functions.}

\item{validations}{array of validation statements}

\item{attachment}{images and other attachments}

\item{layout}{layout schema}

\item{verbose}{should a pretty printed json be shown after each call?}
}
\value{
returns the entire form object with the new field appended.
}
\description{
Create a form field, often a question in the form.
}
\section{Type: Valid Values}{
"calendly" "datedropdown" "email" "file_upload" "group" "legal" "long_text matrix"
"multiple_choice" "nps" "number" "opinion_scale" "payment" "phone_number"
"picture_choice" "ranking" "rating" "short_text" "statement" "website" "yes_no"
}

